+++ backend/src/service/asset_list.rs
@@ -4,8 +4,8 @@
 ("codemirror.min.css", include_bytes!(r#"..\resource\asset/codemirror.min.css.gz"#), "text/css"),
 ("editor.html", include_bytes!(r#"..\resource\asset/editor.html.gz"#), "text/html; charset=utf-8"),
 ("fontawesome.min-5e9e696c59c57e83.css", include_bytes!(r#"..\resource\asset/fontawesome.min-5e9e696c59c57e83.css.gz"#), "text/css"),
-("index-b526a6f2685c5051.js", include_bytes!(r#"..\resource\asset/index-b526a6f2685c5051.js.gz"#), "text/javascript"),
-("index-b526a6f2685c5051_bg.wasm", include_bytes!(r#"..\resource\asset/index-b526a6f2685c5051_bg.wasm.gz"#), "application/wasm"),
+("index-8400e80c80322da0.js", include_bytes!(r#"..\resource\asset/index-8400e80c80322da0.js.gz"#), "text/javascript"),
+("index-8400e80c80322da0_bg.wasm", include_bytes!(r#"..\resource\asset/index-8400e80c80322da0_bg.wasm.gz"#), "application/wasm"),
 ("index-ec4ef66164c33b9c.css", include_bytes!(r#"..\resource\asset/index-ec4ef66164c33b9c.css.gz"#), "text/css"),
 ("index.css", include_bytes!(r#"..\resource\asset/index.css.gz"#), "text/css"),
 ("logo.png", include_bytes!(r#"..\resource\asset/logo.png.gz"#), ""),
