+++ frontend/asset/editor.html
@@ -2,23 +2,29 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title></title>
-    <style>body{padding:0;margin:0}</style>
+    <style>body{margin:1px}</style>
     <link rel="stylesheet" href="/asset/codemirror.min.css" />
     <link rel="stylesheet" href="/asset/toastui-editor.min.css" />
     <script type="application/javascript">
         let editor = null;
         function initEditor() {
+            const c = window.parent.document.getElementById('post-content');
+            const content = c==null?'':c.innerHTML;
             const Editor = toastui.Editor;
             editor = new Editor({
-                el: document.querySelector('#editor'),
+                el: document.querySelector('#editorArea'),
                 previewStyle: 'vertical',
                 initialEditType: 'wysiwyg',
+                initialValue: content,
                 height: '500px',
             });
+        }
+        function getContent() {
+            return editor.getMarkdown();
         }
     </script>
 </head>
 <body>
-<div id="editor">this is editor</div>
+<div id="editorArea"></div>
 <script src="/asset/toastui-editor-all.min.js" onload="initEditor()"></script>
 </body></html>
