+++ frontend/src/component/unauthorized.rs
@@ -27,7 +27,7 @@ impl Component for Unauthorized {
         let mut url = String::from("/management?.redirect_url=");
         url.push_str(&redirect_url);
         html! {
-            <section class="hero is-danger is-bold is-large">
+            <section class="hero is-danger is-bold is-medium">
                 <div class="hero-body">
                     <div class="container">
                         <h1 class="title">
